Agrega apis para registro de prestamos, obtener plazos de pagos en meses y obtencion de clientes.

+++ src/main/java/com/jdreyes/webapi/prestamos/model/dao/PrestamoDAO.java
@@ -1,10 +1,9 @@
 package com.jdreyes.webapi.prestamos.model.dao;
 
-import com.jdreyes.persistence.BaseDAO;
 import com.jdreyes.persistence.BaseQueryExcecutor;
 
 import java.math.BigDecimal;
 
 public interface PrestamoDAO extends BaseQueryExcecutor {
-    void save(Integer idCLiente, String fechaApertura, BigDecimal monto, Integer tasaInteres, Short plazo, String codigo, int ...dias);
+    void save(Integer idCLiente, String fechaApertura, BigDecimal monto, Integer tasaInteres, BigDecimal plazo, String codigo, int ...dias);
 }
